fix11

+++ modules/task_1/gordey_m_matrix_mult/matrix_mult.h
@@ -59,7 +59,7 @@ class SparseM {
       out << "cols: " << m.cols << "\ncol_idx: ";
       for (size_t i = 0; i < m.col_idx.size(); i++)out << m.col_idx[i] << " ";
       out << "\nrows: ";
-      for (int i = 0; i < m.row.size(); i++)out << m.row[i] << " ";
+      for (size_t i = 0; i < m.row.size(); i++)out << m.row[i] << " ";
       out << "\nvalues: ";
       for (size_t i = 0; i < m.value.size(); i++)out << m.value[i] << " ";
       out << "\n";
